+++ apps/client/src/app/(auth)/products/(products)/layout.tsx
@@ -1,0 +1,20 @@
+"use client"
+import Suspense from "@/components/suspense"
+import { GlobalContext } from "@/context/providers/globalstate.provider"
+import { ReactNode, useContext } from "react"
+import Link from "next/link"
+import { ArrowRightIcon } from "@radix-ui/react-icons"
+
+export default function ProductLayout({ children }: { children: ReactNode }) {
+  const [{ userState }] = useContext(GlobalContext)
+
+  return (
+    <Suspense condition={!userState.hasActiveSubscription} fallback={children}>
+      <div className="box">
+        <p className="branding">Hold On</p>
+        <p className="muted-text">Seems like you are not having an active subscription to use/view this Product</p>
+        <Link className="btn btn-primary btn-block" href={"/subscription/plans"}>Subscribe<ArrowRightIcon className="icon-right" /></Link>
+      </div>
+    </Suspense>
+  )
+}