+++ apps/client/src/app/(auth)/subscription/pay/pay-wall.tsx
@@ -6,7 +6,7 @@ export default function PayWall() {
     <div className="box">
       <p className="branding">Hold On</p>
       <p className="muted-text">Seems like you are already having an active subscription</p>
-      <Link className="btn btn-primary btn-block" href="/dashboard">Go to Dashboard<ArrowRightIcon className="icon-right" /></Link>
+      <Link className="btn btn-primary btn-block" href="/products">Go to Products Page<ArrowRightIcon className="icon-right" /></Link>
     </div>
   )
 }