+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Deprecation of platform module
 * Integrated product config in products module
 * Integrated identity-guard component in auth layout
+* Deprecation of dashboard page
+* Introduction of new products & exploreproducts page
 
 ### v10.0.0
 * Migrated entire backend to DDD with CQRS pattern